Fix remaining signed/unsigned warnings * This seems to fix the 64 bit and signed/unsigned warnings in GWEN for   GCC and clang (Xcode). * size_t cast to 32 bit unsigned int. GWEN assumes 32 bit compilation.   This is good enough. GWEN would need a fair bit of work to make 64 compliant.

+++ gwen/src/Controls/Base.cpp
@@ -331,7 +331,7 @@ void Base::RemoveAllChildren()
 unsigned int Base::NumChildren()
 {
     // Include m_InnerPanel's children here?
-    return Children.size();
+    return (unsigned int)Children.size();
 }
 
 Controls::Base* Base::GetChild(unsigned int i)
